Much normalization with develop (#701)
* Brings CMake updates from develop

* Brings reduction in pedantic casts from develop

* Purges UFAIL from the library (#637)

* Committing clang-format changes

* Purges UFAIL from the library

* H5HL_insert change requested in PR


* Removes gratuitous (double)x.yF casts (#632)

* Committing clang-format changes

* Removes gratuitous (double)x.yF casts

* Committing clang-format changes


* Committing clang-format changes

* Cleans up a const warning left over from previous constification (#633)

* Committing clang-format changes

* Adds consts to a few global variables

* Cleans up a const warning left over from previous constification


* Formatted source

* Bring over some VOL registration changes

* VOL cleanup

* H5VL_request_status_t substitutions

* H5F.c cleanup

* Minor API tweaks from develop

* Moves H5G package init/teardown to H5Gint.c

* H5G cleanup

* H5M cleanup

* H5SM cleanup

* H5T cleanup

* H5R cleanup

* H5Lpublic.h cleanup

* H5L cleanup

* H5O cleanup

* H5A, H5CS, and H5AC cleanup

* Moved H5A init/teardown code to H5Aint.c

* Moves H5D I/O functions to H5D.c

* H5D cleanup

* Misc minor cleanup

* H5P close cleanup

* H5Tpublic.h cleanup

* Fixes err_compat test

* H5PLpublic.h cleanup

* Updates H5Ppublic.h

* H5Fpublic.h updates

* H5A.c cleanup

* Brings over H5Aexists and related changes

* Brings CMake shell testing changes from develop

* Close callback changes

* H5R and H5Tcommit normalization

* err_compat test works now

* H5O tweaks

* Updates VOL registration code

* Brings over H5VL_create_object

* H5Tconv.c reformatting

* H5T.c tweaks

* Brings datatype and reference updates from develop

* Brings VOL plugin loading changes from develop

* Brings event sets from develop

* Brings async functions over

* Tools changes

* Brings over many tools changes from develop

* Brings VOL flags from develop

* Fixes h5dump double/float tests

* Updates h5repack tests

* Brings h5diff test changes from develop

* Last h5dump changes

* Brings test changes from develop

* Committing clang-format changes

* Tidied h5_testing()

* Brings chunk iteration code + misc from develop

* Updates vds test

* Enables external link parallel test

* Brings updated property lists from develop

* H5G changes from develop

* H5MF cleanup

* Brings vfd_swmr test back into CMake

* Updates threadsafe test

* Updates plist test

* Brings recent changes from develop

+++ src/H5C.c
@@ -3037,7 +3037,7 @@ H5C_set_cache_auto_resize_config(H5C_t *cache_ptr, H5C_auto_size_ctl_t *config_p
             break;
 
         case H5C_incr__threshold:
-            if ((config_ptr->lower_hr_threshold <= (double)0.0f) || (config_ptr->increment <= (double)1.0f) ||
+            if ((config_ptr->lower_hr_threshold <= 0.0) || (config_ptr->increment <= 1.0) ||
                 ((config_ptr->apply_max_increment) && (config_ptr->max_increment <= 0)))
                 cache_ptr->size_increase_possible = FALSE;
             break;
@@ -3057,21 +3057,21 @@ H5C_set_cache_auto_resize_config(H5C_t *cache_ptr, H5C_auto_size_ctl_t *config_p
             break;
 
         case H5C_decr__threshold:
-            if ((config_ptr->upper_hr_threshold >= (double)1.0f) || (config_ptr->decrement >= (double)1.0f) ||
+            if ((config_ptr->upper_hr_threshold >= 1.0) || (config_ptr->decrement >= 1.0) ||
                 ((config_ptr->apply_max_decrement) && (config_ptr->max_decrement <= 0)))
                 cache_ptr->size_decrease_possible = FALSE;
             break;
 
         case H5C_decr__age_out:
-            if (((config_ptr->apply_empty_reserve) && (config_ptr->empty_reserve >= (double)1.0f)) ||
+            if (((config_ptr->apply_empty_reserve) && (config_ptr->empty_reserve >= 1.0)) ||
                 ((config_ptr->apply_max_decrement) && (config_ptr->max_decrement <= 0)))
                 cache_ptr->size_decrease_possible = FALSE;
             break;
 
         case H5C_decr__age_out_with_threshold:
-            if (((config_ptr->apply_empty_reserve) && (config_ptr->empty_reserve >= (double)1.0f)) ||
+            if (((config_ptr->apply_empty_reserve) && (config_ptr->empty_reserve >= 1.0)) ||
                 ((config_ptr->apply_max_decrement) && (config_ptr->max_decrement <= 0)) ||
-                (config_ptr->upper_hr_threshold >= (double)1.0f))
+                (config_ptr->upper_hr_threshold >= 1.0))
                 cache_ptr->size_decrease_possible = FALSE;
             break;
 
@@ -3793,8 +3793,7 @@ done:
                 HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                             "initial_size must be in the interval [min_size, max_size]")
 
-            if ((config_ptr->min_clean_fraction < (double)0.0f) ||
-                (config_ptr->min_clean_fraction > (double)1.0f))
+            if ((config_ptr->min_clean_fraction < 0.0) || (config_ptr->min_clean_fraction > 1.0))
                 HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                             "min_clean_fraction must be in the interval [0.0, 1.0]")
 
@@ -3810,12 +3809,11 @@ done:
                 HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL, "Invalid incr_mode")
 
             if (config_ptr->incr_mode == H5C_incr__threshold) {
-                if ((config_ptr->lower_hr_threshold < (double)0.0f) ||
-                    (config_ptr->lower_hr_threshold > (double)1.0f))
+                if ((config_ptr->lower_hr_threshold < 0.0) || (config_ptr->lower_hr_threshold > 1.0))
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                 "lower_hr_threshold must be in the range [0.0, 1.0]")
 
-                if (config_ptr->increment < (double)1.0f)
+                if (config_ptr->increment < 1.0)
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                 "increment must be greater than or equal to 1.0")
 
@@ -3830,12 +3828,10 @@ done:
                     break;
 
                 case H5C_flash_incr__add_space:
-                    if ((config_ptr->flash_multiple < (double)0.1f) ||
-                        (config_ptr->flash_multiple > (double)10.0f))
+                    if ((config_ptr->flash_multiple < 0.1) || (config_ptr->flash_multiple > 10.0))
                         HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                     "flash_multiple must be in the range [0.1, 10.0]")
-                    if ((config_ptr->flash_threshold < (double)0.1f) ||
-                        (config_ptr->flash_threshold > (double)1.0f))
+                    if ((config_ptr->flash_threshold < 0.1) || (config_ptr->flash_threshold > 1.0))
                         HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                     "flash_threshold must be in the range [0.1, 1.0]")
                     break;
@@ -3856,10 +3852,10 @@ done:
             }
 
             if (config_ptr->decr_mode == H5C_decr__threshold) {
-                if (config_ptr->upper_hr_threshold > (double)1.0f)
+                if (config_ptr->upper_hr_threshold > 1.0)
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL, "upper_hr_threshold must be <= 1.0")
 
-                if ((config_ptr->decrement > (double)1.0f) || (config_ptr->decrement < (double)0.0f))
+                if ((config_ptr->decrement > 1.0) || (config_ptr->decrement < 0.0))
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL, "decrement must be in the interval [0.0, 1.0]")
 
                 /* no need to check max_decrement as it is a size_t
@@ -3875,8 +3871,8 @@ done:
                 if (config_ptr->epochs_before_eviction > H5C__MAX_EPOCH_MARKERS)
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL, "epochs_before_eviction too big")
 
-                if ((config_ptr->apply_empty_reserve) && ((config_ptr->empty_reserve > (double)1.0f) ||
-                                                          (config_ptr->empty_reserve < (double)0.0f)))
+                if ((config_ptr->apply_empty_reserve) &&
+                    ((config_ptr->empty_reserve > 1.0) || (config_ptr->empty_reserve < 0.0)))
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                 "empty_reserve must be in the interval [0.0, 1.0]")
 
@@ -3886,8 +3882,7 @@ done:
             } /* H5C_decr__age_out || H5C_decr__age_out_with_threshold */
 
             if (config_ptr->decr_mode == H5C_decr__age_out_with_threshold) {
-                if ((config_ptr->upper_hr_threshold > (double)1.0f) ||
-                    (config_ptr->upper_hr_threshold < (double)0.0f))
+                if ((config_ptr->upper_hr_threshold > 1.0) || (config_ptr->upper_hr_threshold < 0.0))
                     HGOTO_ERROR(H5E_ARGS, H5E_BADVALUE, FAIL,
                                 "upper_hr_threshold must be in the interval [0.0, 1.0]")
             } /* H5C_decr__age_out_with_threshold */
@@ -4352,8 +4347,8 @@ done:
         HDassert(cache_ptr);
         HDassert(cache_ptr->magic == H5C__H5C_T_MAGIC);
         HDassert(cache_ptr->cache_accesses >= (cache_ptr->resize_ctl).epoch_length);
-        HDassert((double)0.0f <= (cache_ptr->resize_ctl).min_clean_fraction);
-        HDassert((cache_ptr->resize_ctl).min_clean_fraction <= (double)100.0f);
+        HDassert(0.0 <= (cache_ptr->resize_ctl).min_clean_fraction);
+        HDassert((cache_ptr->resize_ctl).min_clean_fraction <= 100.0);
 
         /* check to see if cache_ptr->resize_in_progress is TRUE.  If it, this
          * is a re-entrant call via a client callback called in the resize
@@ -4376,7 +4371,7 @@ done:
         if (H5C_get_cache_hit_rate(cache_ptr, &hit_rate) != SUCCEED)
             HGOTO_ERROR(H5E_CACHE, H5E_SYSTEM, FAIL, "Can't get hit rate")
 
-        HDassert(((double)0.0f <= hit_rate) && (hit_rate <= (double)1.0f));
+        HDassert((0.0 <= hit_rate) && (hit_rate <= 1.0));
 
         switch ((cache_ptr->resize_ctl).incr_mode) {
             case H5C_incr__off:
